Move gzip header parsing out of bufread module
Header parsing is used by both bufread and write modules.
